requirements.txt
requests>=2.25.1
Pillow>=8.0.0
matplotlib>=3.3.0
numpy>=1.19.0 